requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-crispy-forms==1.8.1
django-formtools==2.2
django-haystack==2.8.1
django-import-export==2.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-redis-cache==2.1.0
django-redis-sessions==0.6.1
django-reversion==3.0.5
django-simple-captcha==0.5.12
django-tinymce==2.8.0
et-xmlfile==1.0.1
future==0.18.2
gedit==0.0.2
httplib2==0.15.0
idna==2.8
jdcal==1.4.1
jieba==0.39
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.7
uWSGI==2.0.18
Whoosh==2.7.4
xlrd==1.2.0
XlsxWriter==1.2.6
xlwt==1.3.0
django-simple-captcha
django-pure-pagination
